requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.4
django-phonenumber-field==7.1.0
django-summernote==0.8.20.0
six==1.16.0
sqlparse==0.4.4
webencodings==0.5.1
